Revise execute package structure
