+++ shardingsphere-governance/shardingsphere-governance-core/shardingsphere-governance-core-context/src/test/resources/logback-test.xml
@@ -25,7 +25,7 @@
     <logger name="org.apache.shardingsphere" level="warn" additivity="false">
         <appender-ref ref="console" />
     </logger>
-    <logger name="org.apache.shardingsphere.infra.executor.sql.execute.driver.jdbc.executor.ExecutorExceptionHandler" level="off" />
+    <logger name="org.apache.shardingsphere.infra.executor.sql.execute.driver.jdbc.ExecutorExceptionHandler" level="off" />
     
     <root>
         <level value="error" />
