+++ shardingsphere-infra/shardingsphere-infra-executor/src/main/java/org/apache/shardingsphere/infra/executor/sql/execute/raw/result/ExecuteResult.java
@@ -15,37 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.executor.sql.execute.raw.execute.result.query;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+package org.apache.shardingsphere.infra.executor.sql.execute.raw.result;
 
 /**
- * Query header.
+ * Execute result.
  */
-@RequiredArgsConstructor
-@Getter
-public final class QueryHeader {
-    
-    private final String schema;
-    
-    private final String table;
-    
-    private final String columnLabel;
-    
-    private final String columnName;
-    
-    private final int columnLength;
-    
-    private final int columnType;
-    
-    private final int decimals;
-    
-    private final boolean signed;
-    
-    private final boolean primaryKey;
-    
-    private final boolean notNull;
-    
-    private final boolean autoIncrement;
+public interface ExecuteResult {
 }
