+++ shardingsphere-infra/shardingsphere-infra-executor/src/test/java/org/apache/shardingsphere/infra/executor/sql/execute/driver/jdbc/ExecutorExceptionHandlerTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.executor.sql.execute.driver.jdbc.executor;
+package org.apache.shardingsphere.infra.executor.sql.execute.driver.jdbc;
 
 import org.junit.After;
 import org.junit.Test;
