+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/communication/jdbc/execute/engine/jdbc/ProxySQLExecutorCallback.java
@@ -20,11 +20,11 @@ package org.apache.shardingsphere.proxy.backend.communication.jdbc.execute.engin
 import org.apache.shardingsphere.infra.database.type.DatabaseType;
 import org.apache.shardingsphere.infra.executor.sql.ConnectionMode;
 import org.apache.shardingsphere.infra.executor.sql.result.QueryResult;
-import org.apache.shardingsphere.infra.executor.sql.execute.raw.execute.result.ExecuteResult;
-import org.apache.shardingsphere.infra.executor.sql.execute.raw.execute.result.query.ExecuteQueryResult;
-import org.apache.shardingsphere.infra.executor.sql.execute.raw.execute.result.query.QueryHeader;
-import org.apache.shardingsphere.infra.executor.sql.execute.raw.execute.result.update.ExecuteUpdateResult;
-import org.apache.shardingsphere.infra.executor.sql.execute.driver.jdbc.executor.impl.DefaultSQLExecutorCallback;
+import org.apache.shardingsphere.infra.executor.sql.execute.raw.result.ExecuteResult;
+import org.apache.shardingsphere.infra.executor.sql.execute.raw.result.query.ExecuteQueryResult;
+import org.apache.shardingsphere.infra.executor.sql.execute.raw.result.query.QueryHeader;
+import org.apache.shardingsphere.infra.executor.sql.execute.raw.result.update.ExecuteUpdateResult;
+import org.apache.shardingsphere.infra.executor.sql.execute.driver.jdbc.impl.DefaultSQLExecutorCallback;
 import org.apache.shardingsphere.infra.executor.sql.result.jdbc.MemoryJDBCQueryResult;
 import org.apache.shardingsphere.infra.executor.sql.result.jdbc.StreamJDBCQueryResult;
 import org.apache.shardingsphere.proxy.backend.communication.jdbc.connection.BackendConnection;
