+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/text/admin/ShowTablesBackendHandler.java
@@ -23,7 +23,7 @@ import org.apache.shardingsphere.infra.executor.sql.result.raw.QueryResultRow;
 import org.apache.shardingsphere.infra.executor.sql.result.raw.RawQueryResult;
 import org.apache.shardingsphere.infra.executor.sql.result.raw.metadata.QueryResultMetaData;
 import org.apache.shardingsphere.infra.executor.sql.result.raw.metadata.QueryResultRowMetaData;
-import org.apache.shardingsphere.infra.executor.sql.execute.raw.execute.result.query.QueryHeader;
+import org.apache.shardingsphere.infra.executor.sql.execute.raw.result.query.QueryHeader;
 import org.apache.shardingsphere.proxy.backend.communication.jdbc.connection.BackendConnection;
 import org.apache.shardingsphere.proxy.backend.context.ProxyContext;
 import org.apache.shardingsphere.proxy.backend.response.BackendResponse;
